Simplify setup
- drop README-PROD.md, prod.sh and docker-compose.prod.yml as these are
  not used for production anymore
- update versions in README-dev.md
- move app-prod.Dockerfile to Dockerfile
- unify settings from devday project devday/settings module
- rename get_variable_cascade to get_setting and implement fallback from
  Vault to environment variables to enable easier local development
- simplify run.sh
- simplify the docker-compose setup for development to only use two
  containers (db and app)
- drop Vault dependency from docker-compose setup
- build a minified production image that can run with read-only rootfs,
  has a high uid for the application user
- exclude coverage information from docker build context
- fix psycopg2 version to avoid database connection issue that happens
  due to an incompatibility between psycopg >= 2.9 and Django < 3 (see
  https://github.com/psycopg/psycopg2/issues/1293)
- move django-debug-toolbar into regular dependencies to allow debugging
  with a prod image if DEBUG is set to true
