+++ .dockerignore
@@ -14,6 +14,8 @@
 **/__pycache__/
 
 /backup/
+/devday/.coverage/
+/devday/coverage-report/
 /devday/media
 /devday/static
 /media/
