+++ Pipfile
@@ -7,7 +7,7 @@ name = "pypi"
 django = "<3"
 django-cms = "<3.9"
 requests = "*"
-psycopg2-binary = "*"
+psycopg2-binary = "<2.9"
 python-magic = "*"
 pillow = "*"
 django-crispy-forms = "*"
@@ -36,11 +36,11 @@ django-postgres-extra = "*"
 djangorestframework = "*"
 djangocms-column = "*"
 django-treebeard = "<4.5"
+django-debug-toolbar = "*"
 
 [dev-packages]
 coverage = "*"
 coveralls = "*"
-django-debug-toolbar = "*"
 dj-inmemorystorage = "*"
 mock = "*"
 pip = "==18.0"
