--- docker-compose.dev.yml
@@ -1,43 +1,0 @@
-version: '3'
-services:
-  db:
-    image: postgres:12
-    ports:
-      - "15432:5432"
-    env_file:
-      dev-env-db
-    volumes:
-      - "pg_data:/var/lib/postgresql/data"
-      - "./docker/db/devday_db.sql:/docker-entrypoint-initdb.d/init-user-db.sql"
-  vault:
-    image: vault:1.6.0
-    environment:
-      VAULT_DEV_ROOT_TOKEN_ID: devday_root
-      VAULT_DEV_LISTEN_ADDRESS: 0.0.0.0:8200
-    command: "server -dev"
-  app:
-    image: ${DOCKER_REGISTRY}devday_website_app_dev:latest
-    build:
-      context: .
-      dockerfile: app-dev.Dockerfile
-      args:
-        http_proxy: ${http_proxy}
-        no_proxy: ${no_proxy},vault
-    ports:
-      - "8000:8000"
-    environment:
-      DEVDAY_PG_DBNAME: devday
-      DEVDAY_PG_HOST: db
-      DEVDAY_PG_PORT: 5432
-      DEVDAY_PG_USER: devday
-      DJANGO_SETTINGS_MODULE: devday.settings.dev
-      VAULT_URL: http://vault:8200
-    env_file:
-      dev-env
-    volumes:
-      - "devday_media:/app/media"
-      - "devday_static:/app/static"
-      - "./devday:/app"
-
-volumes:
-  pg_data:
