+++ docker/app/start-application.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
-set -e
+set -eu
 
-python3 manage.py collectstatic --noinput
 python3 manage.py migrate --noinput
+python3 manage.py collectstatic --noinput
 uwsgi --ini /app/uwsgi.ini
